refactor: handle exceptions and standard format

+++ app/api/endpoints/crud_api.py
@@ -1,14 +1,21 @@
 from typing import Optional
-from fastapi import APIRouter,Form, File, UploadFile
+from fastapi import APIRouter,Form, File, UploadFile,HTTPException
 from app.crud import crud_user_data
 from utils.dbconnection import collection_name
+from utils.response import response
 import os
 from bson.objectid import ObjectId
+from fastapi_pagination import Page, paginate
+from fastapi_pagination.utils import disable_installed_extensions_check
+
+disable_installed_extensions_check()
 
 # Define the directory to save uploaded files
 UPLOAD_DIRECTORY = "uploaded_files"
 
 router = APIRouter()
+
+accepted_file_type = ["image/png","image/jpeg","image/jpg"]
 
 @router.post("/create_api")
 async def create_api(
@@ -17,6 +24,12 @@ async def create_api(
     image: UploadFile = File(...),
 ):
     try:
+        if image.content_type not in accepted_file_type:
+            return response(1003)
+        
+        if os.path.exists(f"{UPLOAD_DIRECTORY}/{image.filename}"):
+            return response(1009)
+
         file_location = os.path.join(UPLOAD_DIRECTORY, image.filename)
         with open(file_location, "wb") as file_object:
             file_object.write(await image.read())
@@ -28,18 +41,20 @@ async def create_api(
         }
         user = crud_user_data.create(collection_name,user_data)
         if user:
-            return {
-                "Message":"User Created Successfully",
-                "Data":user
-            }
+            return response(1000,user)
         
-    except Exception as e:
-        raise e
+    except Exception:
+        raise response(1005)
 
-@router.post("/read_api")
-def read_api():
-    users_data = crud_user_data.read(collection_name)
-    return users_data
+@router.get("/read_api")
+def read_api() -> Page:
+    try:
+        users_data = crud_user_data.read(collection_name)
+
+        return response(1004,paginate(users_data).items)
+
+    except Exception:
+        return response(1005)
 
 @router.post("/update_api")
 async def update_api(
@@ -48,31 +63,43 @@ async def update_api(
     summery: Optional[str] = Form(None),
     image: Optional[UploadFile] = File(None)
 ):
-    if image:
-        file_location = os.path.join(UPLOAD_DIRECTORY, image.filename)
-        with open(file_location, "wb") as file_object:
-            file_object.write(await image.read())
-    if id:
-        data = collection_name.find_one({"_id":ObjectId(id)})
+    try:
+        if image:
+            if image.content_type not in accepted_file_type:
+                return response(1003)
 
-    user_data={
-        "name":name if name else data['name'],
-        "image":f"files/{image.filename}" if image else data['image'],
-        "summery":summery if summery else data['summery']
-    }
-    update_data = crud_user_data.update(collection_name,id,user_data)
+            if os.path.exists(f"{UPLOAD_DIRECTORY}/{image.filename}"):
+                return response(1009)
 
-    if update_data:
-        return {
-            "message": "Data Update Successfully"
+            file_location = os.path.join(UPLOAD_DIRECTORY, image.filename)
+            with open(file_location, "wb") as file_object:
+                file_object.write(await image.read())
+        if id:
+            data = collection_name.find_one({"_id":ObjectId(id)})
+
+        user_data={
+            "name":name if name else data['name'],
+            "image":f"files/{image.filename}" if image else data['image'],
+            "summery":summery if summery else data['summery']
         }
+        update_data = crud_user_data.update(collection_name,id,user_data)
+
+        if update_data:
+            return response(1006)
+
+    except Exception:
+        return response(1005)
 
 @router.post("/delete_api")
 def delete_api(
     id:str = Form()
 ):
-
-    users_data = crud_user_data.delete(collection_name,id)
-    if users_data:
-        return {"Message": "Data Deleted Successfully."}
-    return users_data
+    try:
+        users_data = crud_user_data.delete(collection_name,id)
+        if not users_data:
+            return response(1008)
+        
+        return response(1007)
+    
+    except Exception:
+        return response(1005)
